refactor(goesdr): add latlon module
Add latitude and longitude grid data extraction module.
